+++ be/src/vec/core/block.cpp
@@ -819,113 +819,6 @@ inline bool Block::is_column_data_null(const doris::TypeDescriptor& type_desc,
     }
 }
 
-// TODO: need to refactor this function, too long.
-void Block::deep_copy_slot(void* dst, MemPool* pool, const doris::TypeDescriptor& type_desc,
-                           const StringRef& data_ref, const IColumn* column, int row,
-                           bool padding_char) {
-    if (type_desc.is_collection_type()) {
-        if (type_desc.type != TYPE_ARRAY) {
-            return;
-        }
-
-        Field field;
-        column->get(row, field);
-        const auto& array = field.get<Array>();
-        auto collection_value = reinterpret_cast<CollectionValue*>(dst);
-        auto item_type_desc = type_desc.children.front();
-        CollectionValue::init_collection(pool, array.size(), item_type_desc.type, collection_value);
-
-        const ColumnArray* array_column = nullptr;
-        if (is_column_nullable(*column)) {
-            auto& nested_column =
-                    reinterpret_cast<const ColumnNullable*>(column)->get_nested_column();
-            array_column = reinterpret_cast<const ColumnArray*>(&nested_column);
-        } else {
-            array_column = reinterpret_cast<const ColumnArray*>(column);
-        }
-        auto item_column = array_column->get_data_ptr().get();
-        auto offset = array_column->get_offsets()[row - 1];
-        auto iterator = collection_value->iterator(item_type_desc.type);
-        for (int i = 0; i < collection_value->length(); ++i) {
-            if (array[i].is_null()) {
-                const auto& null_value = doris_udf::AnyVal(true);
-                iterator.set(&null_value);
-            } else {
-                auto item_offset = offset + i;
-                const auto& data_ref = item_type_desc.type != TYPE_ARRAY
-                                               ? item_column->get_data_at(item_offset)
-                                               : StringRef {};
-                if (item_type_desc.is_date_type()) {
-                    // In CollectionValue, date type data is stored as either uint24_t or uint64_t.
-                    DateTimeValue datetime_value;
-                    deep_copy_slot(&datetime_value, pool, item_type_desc, data_ref, item_column,
-                                   item_offset, padding_char);
-                    DateTimeVal datetime_val;
-                    datetime_value.to_datetime_val(&datetime_val);
-                    iterator.set(&datetime_val);
-                } else if (item_type_desc.is_decimal_v2_type()) {
-                    // In CollectionValue, decimal type data is stored as decimal12_t.
-                    DecimalV2Value decimal_value;
-                    deep_copy_slot(&decimal_value, pool, item_type_desc, data_ref, item_column,
-                                   item_offset, padding_char);
-                    DecimalV2Val decimal_val;
-                    decimal_value.to_decimal_val(&decimal_val);
-                    iterator.set(&decimal_val);
-                } else {
-                    deep_copy_slot(iterator.get(), pool, item_type_desc, data_ref, item_column,
-                                   item_offset, padding_char);
-                }
-            }
-            iterator.next();
-        }
-    } else if (type_desc.is_date_type()) {
-        VecDateTimeValue ts =
-                *reinterpret_cast<const doris::vectorized::VecDateTimeValue*>(data_ref.data);
-        DateTimeValue dt;
-        ts.convert_vec_dt_to_dt(&dt);
-        memcpy(dst, &dt, sizeof(DateTimeValue));
-    } else if (type_desc.type == TYPE_OBJECT) {
-        auto bitmap_value = (BitmapValue*)(data_ref.data);
-        auto size = bitmap_value->getSizeInBytes();
-
-        // serialize the content of string
-        // TODO: NEED TO REWRITE COMPLETELY. the way writing now is WRONG.
-        // StringRef shouldn't managing exclusive memory cause it will break RAII.
-        // besides, accessing object which is essentially const by non-const object
-        // is UB!
-        auto string_slot = reinterpret_cast<StringRef*>(dst);
-        string_slot->data = reinterpret_cast<char*>(pool->allocate(size));
-        bitmap_value->write(const_cast<char*>(string_slot->data)); //!
-        string_slot->size = size;
-    } else if (type_desc.type == TYPE_HLL) {
-        auto hll_value = (HyperLogLog*)(data_ref.data);
-        auto size = hll_value->max_serialized_size();
-        auto string_slot = reinterpret_cast<StringRef*>(dst);
-        string_slot->data = reinterpret_cast<char*>(pool->allocate(size));
-        size_t actual_size = hll_value->serialize((uint8_t*)string_slot->data);
-        string_slot->size = actual_size;
-    } else if (type_desc.is_string_type()) { // TYPE_OBJECT and TYPE_HLL must be handled before.
-        memcpy(dst, (const void*)(&data_ref), sizeof(data_ref));
-        // Copy the content of string
-        if (padding_char && type_desc.type == TYPE_CHAR) {
-            // serialize the content of string
-            auto string_slot = reinterpret_cast<StringRef*>(dst);
-            string_slot->data = reinterpret_cast<char*>(pool->allocate(type_desc.len));
-            string_slot->size = type_desc.len;
-            memset(const_cast<char*>(string_slot->data), 0, type_desc.len);             //!
-            memcpy(const_cast<char*>(string_slot->data), data_ref.data, data_ref.size); //!
-        } else {
-            auto str_ptr = pool->allocate(data_ref.size);
-            memcpy(str_ptr, data_ref.data, data_ref.size);
-            auto string_slot = reinterpret_cast<StringRef*>(dst);
-            string_slot->data = reinterpret_cast<char*>(str_ptr);
-            string_slot->size = data_ref.size;
-        }
-    } else {
-        memcpy(dst, data_ref.data, data_ref.size);
-    }
-}
-
 MutableBlock::MutableBlock(const std::vector<TupleDescriptor*>& tuple_descs, int reserve_size,
                            bool ignore_trivial_slot) {
     for (auto tuple_desc : tuple_descs) {
